+++ postgres/src/driver/generic.rs
@@ -1,6 +1,6 @@
 use core::{
-    convert::Infallible,
     future::{poll_fn, Future},
+    mem::MaybeUninit,
     pin::Pin,
 };
 
@@ -9,16 +9,21 @@ use std::{collections::VecDeque, io};
 use postgres_protocol::message::backend;
 use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender};
 use xitca_io::{
-    bytes::{BufInterest, BufRead, BufWrite, BytesMut, WriteBuf},
+    bytes::{Buf, BufRead, BytesMut},
     io::{AsyncIo, Interest},
 };
-use xitca_unsafe_collection::futures::{Select as _, SelectOutput};
+use xitca_unsafe_collection::{
+    bytes::BufList,
+    futures::{Select as _, SelectOutput},
+};
 
 use crate::error::Error;
 
 use super::codec::{Request, Response, ResponseMessage, ResponseSender, SenderState};
 
 type PagedBytesMut = xitca_unsafe_collection::bytes::PagedBytesMut<4096>;
+
+const BUF_LIST_CNT: usize = 32;
 
 pub(crate) type DriverRx = UnboundedReceiver<Request>;
 
@@ -47,10 +52,11 @@ impl DriverTx {
 
 pub struct GenericDriver<Io> {
     pub(crate) io: Io,
-    pub(crate) write_buf: WriteBuf,
     pub(crate) read_buf: PagedBytesMut,
     pub(crate) state: DriverState,
     pub(crate) res: VecDeque<ResponseSender>,
+    write_bufs: BufList<BytesMut, BUF_LIST_CNT>,
+    want_flush_io: bool,
 }
 
 pub(crate) enum DriverState {
@@ -77,20 +83,25 @@ where
         (
             Self {
                 io,
-                write_buf: WriteBuf::new(),
                 read_buf: PagedBytesMut::new(),
-                res: VecDeque::new(),
                 state: DriverState::Running(rx),
+                res: VecDeque::new(),
+                write_bufs: BufList::new(),
+                want_flush_io: false,
             },
             DriverTx(tx),
         )
     }
 
+    fn want_write_io(&self) -> bool {
+        !self.write_bufs.is_empty() || self.want_flush_io
+    }
+
     pub(crate) async fn send(&mut self, msg: BytesMut) -> Result<(), Error> {
-        self.write_buf_extend(&msg);
+        self.write_bufs.push(msg);
         loop {
             self.try_write()?;
-            if self.write_buf.is_empty() {
+            if self.write_bufs.is_empty() {
                 return Ok(());
             }
             self.io.ready(Interest::WRITABLE).await?;
@@ -107,14 +118,24 @@ where
                 return Ok(Some(msg));
             }
 
-            let interest = if self.write_buf.want_write_io() {
+            let interest = if self.want_write_io() {
                 Interest::READABLE.add(Interest::WRITABLE)
             } else {
                 Interest::READABLE
             };
 
             let select = match self.state {
-                DriverState::Running(ref mut rx) => rx.recv().select(self.io.ready(interest)).await,
+                DriverState::Running(ref mut rx) => {
+                    async {
+                        if self.write_bufs.is_full() {
+                            core::future::pending().await
+                        } else {
+                            rx.recv().await
+                        }
+                    }
+                    .select(self.io.ready(interest))
+                    .await
+                }
                 DriverState::Closing(ref mut e) => {
                     if !interest.is_writable() && self.res.is_empty() {
                         // no interest to write to io and all response have been finished so
@@ -131,8 +152,9 @@ where
             match select {
                 // batch message and keep polling.
                 SelectOutput::A(Some(req)) => {
-                    self.write_buf_extend(req.msg.as_ref());
                     self.res.push_back(req.tx);
+                    self.write_bufs.push(req.msg);
+                    self.want_flush_io = false;
                 }
                 SelectOutput::B(ready) => {
                     let ready = ready?;
@@ -144,7 +166,8 @@ where
                             // when write error occur the driver would go into half close state(read only).
                             // clearing write_buf would drop all pending requests in it and hint the driver
                             // no future Interest::WRITABLE should be passed to AsyncIo::ready method.
-                            self.write_buf.clear();
+                            self.write_bufs.clear();
+                            self.want_flush_io = false;
 
                             // enter closed state and no more request would be received from channel.
                             // requests inside it would eventually be dropped after shutdown completed.
@@ -170,19 +193,37 @@ where
         }
     }
 
-    fn write_buf_extend(&mut self, buf: &[u8]) {
-        let _ = self.write_buf.write_buf(|w| {
-            w.extend_from_slice(buf);
-            Ok::<_, Infallible>(())
-        });
-    }
-
     fn try_read(&mut self) -> Result<(), Error> {
         self.read_buf.do_io(&mut self.io).map_err(Into::into)
     }
 
     fn try_write(&mut self) -> io::Result<()> {
-        self.write_buf.do_io(&mut self.io)
+        loop {
+            if self.want_flush_io {
+                match io::Write::flush(&mut self.io) {
+                    Ok(_) => self.want_flush_io = false,
+                    Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => {}
+                    Err(e) => return Err(e),
+                }
+                break;
+            }
+
+            let mut bufs = [const { MaybeUninit::uninit() }; BUF_LIST_CNT];
+            let slice = self.write_bufs.chunks_vectored_uninit_into_init(&mut bufs);
+            match io::Write::write_vectored(&mut self.io, slice) {
+                Ok(0) => return Err(io::ErrorKind::WriteZero.into()),
+                Ok(n) => {
+                    self.write_bufs.advance(n);
+                    if self.write_bufs.is_empty() {
+                        self.want_flush_io = true;
+                    }
+                }
+                Err(ref e) if e.kind() == io::ErrorKind::WouldBlock => break,
+                Err(e) => return Err(e),
+            }
+        }
+
+        Ok(())
     }
 
     fn try_decode(&mut self) -> Result<Option<backend::Message>, Error> {